Safari-609.1.11.1/LayoutTests/crypto/subtle/hmac-import-jwk-key-export-raw-key-expected.txt

Test importing a JWK HMAC key with SHA-1 and then export it in raw format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
Exporting a key...
PASS Base64URL.stringify(exportedRawKey) is jwkKey.k
PASS successfullyParsed is true

TEST COMPLETE
